requirements.txt
requests==2.32.2
selenium==4.28.0
beautifulsoup4==4.12.3 
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
lxml==4.6.4
soupsieve==2.3.1
urllib3==1.26.7
randomheader==0.4
httpx-ip-rotator==0.1.4
python-dotenv==1.0.1
